zeta0 verification test

+++ mach/mach.cpp
@@ -8,7 +8,7 @@
 
 namespace mach {
 
-    double machElement(int i, double x) {
+    double machElement(long i, double x) {
 
         //fancy but faster way than pow to alternate the the minus, thanks to unit tests :D
         double sign = 1.0 - (2.0 * ((i & 0x1) ^ 0x1));
@@ -17,19 +17,19 @@ namespace mach {
         return sign * pow(x, factor)/factor;
     }
 
-    double singleMach(int n, double x) {
+    double singleMach(long n, double x) {
         assert(n > 0);
         assert(x >= -1.0 && x <= 1.0);
 
 
         double sum = 0.0;
-        for (int i = 1; i <= n; i++){
+        for (long i = 1; i <= n; i++){
             sum += machElement(i, x);
         }
         return sum;
     }
 
-    double getPIMachSingleT(int n) {
+    double getPIMachSingleT(long n) {
         auto series_1 = singleMach(n, 1./5.);
         auto series_2 = singleMach(n, 1./239.);
